requirements.txt
discord.py==2.6.2
python-dotenv==1.1.1
trello==0.9.7.3
mailjet-rest==1.5.1